Added the event handlers for the buttons in the Search Window. Cleaned up a couple of other things.
