+++ GroupProject/GroupProject/Product.xaml
@@ -1,0 +1,16 @@
+﻿<Window x:Class="GroupProject.Product"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:GroupProject"
+        mc:Ignorable="d"
+        Title="Product" Height="471.304" Width="698.261">
+    <Grid>
+        <DataGrid x:Name="dgProduct" HorizontalAlignment="Left" Margin="10,19,0,0" VerticalAlignment="Top" Height="412" Width="567"/>
+        <Button x:Name="btnProdAdd" Content="Add" HorizontalAlignment="Left" Margin="595,75,0,0" VerticalAlignment="Top" Width="75" Click="btnProdAdd_Click"/>
+        <Button x:Name="btnProdEdt" Content="Edit" HorizontalAlignment="Left" Margin="595,179,0,0" VerticalAlignment="Top" Width="75" Click="btnProdEdt_Click"/>
+        <Button x:Name="btnProdDel" Content="Delete" HorizontalAlignment="Left" Margin="595,126,0,0" VerticalAlignment="Top" Width="75" Click="btnProdDel_Click"/>
+
+    </Grid>
+</Window>
