--- GroupProject/GroupProject/Search.xaml
@@ -1,12 +1,0 @@
-﻿<Window x:Class="GroupProject.Search"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:GroupProject"
-        mc:Ignorable="d"
-        Title="Search" Height="300" Width="300">
-    <Grid>
-        
-    </Grid>
-</Window>
