--- GroupProject/GroupProject/Search.xaml.cs
@@ -1,27 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
-
-namespace GroupProject
-{
-    /// <summary>
-    /// Interaction logic for Search.xaml
-    /// </summary>
-    public partial class Search : Window
-    {
-        public Search()
-        {
-            InitializeComponent();
-        }
-    }
-}
